+++ components/test_data/inspirations.ts
@@ -1,0 +1,39 @@
+import type { InspirationEntry } from "../InspirationMgr";
+
+export const inspirationTestData: InspirationEntry[] = [
+  {
+    category: "artist",
+    value: "Rene Magritte",
+    imageId: "6d26c249-1a8e-414c-a800-a5baa34a1b8c",
+    imageUrl: "bfe709ab-6a5e-4626-94d4-9ccb1ecd5c4d-0-2003297801.png",
+    prompt: "beautiful symmetric face, by Rene Magritte",
+  },
+  {
+    category: "?",
+    value: "?",
+    imageId: "90e3cbf4-2658-42d4-a73e-98d962e3e9ed",
+    imageUrl: "914c83b6-eeb4-47ac-8f97-2645e055fb78-0-2003297801.png",
+    prompt: "beautiful symmetric face, by Sonia Delaunay",
+  },
+  {
+    category: "artist",
+    value: "Victto Ngai",
+    imageId: "2ba1f8b6-5f64-4cc5-b610-c89198906508",
+    imageUrl: "5bf22837-7f3a-4993-8b10-82e2c6a37180-0-2003297801.png",
+    prompt: "beautiful symmetric face, by Victto Ngai",
+  },
+  {
+    category: "artist",
+    value: "Andy Warhol",
+    imageId: "1f4df931-abf3-4dd9-8988-6b9d507239ff",
+    imageUrl: "422f4c8d-8c82-47d8-9076-103578fe1235-0-2003297801.png",
+    prompt: "beautiful symmetric face, andy warhol",
+  },
+  {
+    category: "artist",
+    value: "Karol Bak",
+    imageId: "cbcd9acd-9a19-46ab-96f1-31a3e8b39c8e",
+    imageUrl: "8958cfcf-a4f6-4df7-a1b7-ec8264fc7739-0-2003297801.png",
+    prompt: "beautiful symmetric face, karol bak",
+  },
+];